+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
   key that can be set, and will treat newly-found pages as changed,
   and display a diff from the empty string (useful for `diff_tool`
   or `diff_filter` with side effects)
+- New reporter: `discord`
 
 ### Changed